refactor : custom 함수 따로 분리
- custom 함수들을 custom.js 파일에 따로 분리
- util 함수 중에 _$ 라는 커스텀 함수를 이름 지어 같은 역할을 하도록 생성
- 추후 테스트할 예정
